BUILD: updated build script

+++ build.sh
@@ -14,7 +14,7 @@ fi
 
 
 # define toolchain
-CC=${HOME}/android/cm-11.0/prebuilts/gcc/linux-x86/arm/arm-eabi-4.6/bin/arm-eabi-
+CC=${HOME}/android/toolchains/arm-cortex_a15-linux-gnueabihf-linaro_4.8.3-2014.04/bin/arm-cortex_a15-linux-gnueabihf-
 
 # create .config
 if [ ! -e ./.config ]; then
@@ -26,15 +26,17 @@ fi
 # build the kernel
 echo "**** Building kernel ****"
 make -j3 ARCH=arm SUBARCH=arm CROSS_COMPILE=$CC
-echo "**** Successfully built kernel ****"
+if [ -e ./arch/arm/boot/zImage ]; then
+	echo "**** Successfully built kernel ****"
 
-echo "**** Copying kernel to build_result/kernel/ ****"
-mkdir -p ./build_result/kernel/
-cp ./arch/arm/boot/zImage ./build_result/kernel/zImage
+	echo "**** Copying kernel to build_result/kernel/ ****"
+	mkdir -p ./build_result/kernel/
+	cp ./arch/arm/boot/zImage ./build_result/kernel/zImage
 
-echo "**** Copying all built modules (.ko) to build_result/modules/ ****"
-mkdir -p ./build_result/modules/
-for file in $(find ./ -name *.ko); do
-	cp $file ./build_result/modules/
-done
+	echo "**** Copying all built modules (.ko) to build_result/modules/ ****"
+	mkdir -p ./build_result/modules/
+	for file in $(find ./ -name *.ko); do
+		cp $file ./build_result/modules/
+	done
+fi
